+++ app/src/main/res/layout/fragment_accept_join_request.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AcceptJoinRequestFragment"
+    tools:context=".presentation.teambuild.AcceptJoinRequestFragment"
     android:background="@color/bg_popup">
