+++ app/src/main/res/layout/fragment_team_build_post_view_general.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TeamBuildPostViewGeneral">
+    tools:context=".presentation.teambuild.TeamBuildPostViewGeneral">
 
     <LinearLayout
         android:id="@+id/ll_tv_subject_name"
