+++ app/src/main/res/layout/fragment_team_build_post_view_leader.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TeamBuildPostViewLeader">
+    tools:context=".presentation.teambuild.TeamBuildPostViewLeader">
 
     <LinearLayout
         android:id="@+id/ll_tv_subject_name"
