+++ app/src/main/res/layout/fragment_team_build_writing.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TeamBuildWritingFragment">
+    tools:context=".presentation.teambuild.TeamBuildWritingFragment">
 
     <LinearLayout
         android:id="@+id/ll_tv_subject_name"
